requirements/requirements_6.txt
absl-py==2.1.0
astunparse==1.6.3
audioread==3.0.1
azure-cognitiveservices-speech==1.41.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
distlib==0.3.8
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
hmmlearn==0.3.2
ibm-cloud-sdk-core==3.21.0
ibm-watson==8.1.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
moviepy==1.0.3
msgpack==1.1.0
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
protobuf==4.25.5
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.9.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.5.0
tesseract==0.1.3
threadpoolctl==3.5.0
tifffile==2024.9.20
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.5
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
